Med: sbd: Fix definition of size_t.

+++ lib/stonith/sbd.c
@@ -53,7 +53,7 @@ static const char *watchdogdev		= "/dev/watchdog";
 static char *	local_uname;
 
 /* Global, non-tunable variables: */
-static unsigned long	sector_size	= 0;
+static size_t	sector_size		= 0;
 static int	watchdogfd 		= -1;
 static int	devfd;
 static char	*devname;
